Change the default font for Windows
Chart::Clicker currently fails to compile on Windows, as the Sans font doesn't exist.  The Chart::Clicker tests add text labels, which use Graphics::Primitive::Font and Graphics::Primitive::Driver::GD to work out the size of the text label bounding box.  This is undefined because the Sans font doesn't exist, so all the tests fail.  This change also reads the GRAPHICS_PRIMITIVE_DEFAULT_FONT environment variable to allow the default font to be overridden